+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - `get_connectivity_html()` returns HTML as non-scalable #3213
 - add update-serial to `DC_EVENT_WEBXDC_STATUS_UPDATE` #3215
 - Speed up message receiving via IMAP a bit #3225
+- mark messages as seen on IMAP in batches #3223
 
 
 ## 1.77.0